+++ Classes/ModelInputFiles/GUT_BatchInputFileBuilder.cs
@@ -10,7 +10,7 @@ namespace CHaMPWorkbench.Classes.ModelInputFiles
     {
         private GUTInputProperties m_Inputs;
 
-        public GUT_BatchInputFileBuilder(string sDBCon, string sBatchName, bool bMakeOnlyBatch, string sMonitoringDataFolder, string sOutputFolder, ref Dictionary<int, string> dVisits, string sInputFileName, GUTInputProperties theInputs)
+        public GUT_BatchInputFileBuilder(string sDBCon, string sBatchName, bool bMakeOnlyBatch, string sMonitoringDataFolder, string sOutputFolder, ref Dictionary<long, string> dVisits, string sInputFileName, GUTInputProperties theInputs)
             : base(CHaMPWorkbench.Properties.Settings.Default.ModelType_GUT, sDBCon, sBatchName, bMakeOnlyBatch, sMonitoringDataFolder, sOutputFolder, ref  dVisits, sInputFileName)
         {
             m_Inputs = new GUTInputProperties(theInputs);
